refactor: remove 'x' option from bash shell settings in entry script
The 'x' option in the 'set -eux' command was causing the shell to print all executed commands before running them in the 'entrypoint.sh' script. This had the potential to disclose sensitive details in logs. For enhanced security and to prevent unnecessary logs, the 'x' option has been removed from the shell settings.