refactor: optimize if conditions of usage calculation in `entrypoint.sh`
This commit streamlines the calculation of GitHub repo usage and the action cancellation in `entrypoint.sh`. The changes make sure the action cancellation is only dealt with for private repos, minimizing unnecessary calculations for public repos. Additionally, usage calculation and exceeding threshold checks now only occur if the repository is not public, contributing to a more efficient overall script operation.